+++ lib/commands/plugin-remove.sh
@@ -6,4 +6,6 @@ plugin_remove_command() {
 
   rm -rf $plugin_path
   rm -rf $(asdf_dir)/installs/${plugin_name}
+
+  grep -l "asdf-plugin: ${plugin_name}" $(asdf_dir)/shims/* 2>/dev/null | xargs rm -f
 }