+++ CHANGELOG.md
@@ -1,11 +1,44 @@
 #Changelog
 
-##0.2.0-dev
+##0.3.1-dev
+
+##0.3.0
 
 Features
 
+* Add `update` command to make it easier to update asdf to the latest release (#172, #180)
+* Add support for `system` version to allow passthrough to system installed tools (#55, #182)
+
+Fixed Bugs
+
+* Set `GREP_OPTIONS` and `GREP_COLORS` variables in util.sh so grep is always invoked with the correct settings (#170)
+* Export `ASDF_DIR` variable so the Zsh plugin can locate asdf if it's in a custom location (#156)
+* Don't add execute permission to files in a plugin's bin directory when adding the plugin (#124, #138, #154)
+
+##0.2.1
+
+Features
+
+* Determine global tool version even when used outside of home directory (#106)
+
+Fixed Bugs
+
+* Correct reading of `ref:` and `path:` versions (#112)
+* Remove shims when uninstalling a version or removing a plugin (#122, #123, #125, #128, #131)
+* Add a helpful error message to the install command (#135)
+
+##0.2.0
+
+Features
+
+* Improve plugin API for legacy file support (#87)
 * Unify `asdf local` and `asdf global` version getters as `asdf current` (#83)
 * Rename `asdf which` to `asdf current` (#78)
+
+Fixed Bugs
+
+* Fix bug that caused the `local` command to crash when the directory contains whitespace (#90)
+* Misc typo corrections (#93, #99)
 
 ##0.1.0